solved rebasing coflicts + updated event-store to match events branch, moved all constants to constants.ts and design enhancement
